add playlist management, css & utils

+++ backend/models/playlist.model.js
@@ -1,5 +1,11 @@
 import dbConnect from '../controllers/db/index.js'
 
+export const find5LatestPlaylists = async () => {
+    const sql_string = 'SELECT playlists.* FROM playlists ORDER BY created_at DESC LIMIT 5'
+
+    return await dbConnect(sql_string)
+}
+
 export const findAllPlaylistsByUserId = async (arg) => {
     const sql_string = 'SELECT playlists.* FROM playlists WHERE user_id = ?'
 
@@ -19,13 +25,13 @@ export const addOnePlaylist = async (arg) => {
 }
 
 export const updateOnePlaylistById = async (arg) => {
-    const sql_string = 'UPDATE playlists SET name = ?, image = ? WHERE id = ?'
+    const sql_string = 'UPDATE playlists SET name = ?, image = ? WHERE id = ? AND user_id = ?'
 
     return await dbConnect(sql_string, arg)
 }
 
 export const deleteOnePlaylistById = async (arg) => {
-    const sql_string = 'DELETE FROM playlists WHERE id = ?'
+    const sql_string = 'DELETE FROM playlists WHERE id = ? AND user_id = ?'
 
     return await dbConnect(sql_string, arg)
 }
